Rename Jupyter to Research add nuget
- Rename Jupyter to Research and add it's nuget package
