+++ Research/Properties/AssemblyInfo.cs
@@ -5,11 +5,11 @@ using System.Runtime.InteropServices;
 // General Information about an assembly is controlled through the following 
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
-[assembly: AssemblyTitle("QuantConnect.Jupyter")]
-[assembly: AssemblyDescription("QuantConnect LEAN Engine: Jupyter Project - Core implementation for jupyter research environment")]
+[assembly: AssemblyTitle("QuantConnect.Research")]
+[assembly: AssemblyDescription("QuantConnect LEAN Engine: Research Project - Core implementation for jupyter research environment")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("QuantConnect.Jupyter")]
+[assembly: AssemblyProduct("QuantConnect.Research")]
 [assembly: AssemblyCopyright("Copyright ©  2017")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
